chore: remove seller blocking
After the new Tori.fi update, there is no way to uniquely identify
the seller of an ad. I've decided to get rid of the seller
blacklist feature all together, despite it still being possible for
Huuto.net ads.

This is an unfortunate design choice on Tori.fi's part
and in my personal opinion the complete opposite of what should've been
done.